Make sources as compatible with gcc 6.5.0 as possible.
 * Don't use <hardware/intbits.h>
 * hardware.h should only depend on types.h
 * Compile flatshade-convex, showpchg and uvlight as well.
 * Avoid warnings about aliasing.
 * Fix compiler claiming that initializer is not constant.
 * Fix section names for startup code.
 * Fix all warnings in effects code.
 * Put mul/div/mod instruction emulation into .text section.
 * Fix pointer signedness issues.
 * Hide console initialization details.
 * Clean up logging.
 * Introduce kvprintf and snprintf in place of RawDoFmt.
 * Fix sintab.py
 * Don't use <exec/types.h> definitions.
 * Modify tmxconv.py to use macro.
 * Enable extra warnings (including -Wstrict-prototypes).

+++ a500/base/include/circle.h
@@ -3,7 +3,7 @@
 
 #include "gfx.h"
 
-__regargs void Circle(BitmapT *bitmap, LONG plane, WORD x0, WORD y0, WORD r);
-__regargs void CircleEdge(BitmapT *bitmap, LONG plane, WORD x0, WORD y0, WORD r);
+__regargs void Circle(BitmapT *bitmap, int plane, short x0, short y0, short r);
+__regargs void CircleEdge(BitmapT *bitmap, int plane, short x0, short y0, short r);
 
 #endif
